[fix] 日付変換ロジックをhooksへ移動、関連する型定義を整理
